consumer/requirements.txt
pika
paho-mqtt
osmnx
shapely 
geopandas